requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
pandas>=2.2.0
playwright>=1.50.0
python-dateutil>=2.9.0
sqlite-utils>=3.38

# PDF text extraction — required for extracting decision text from downloaded PDFs
pdfminer.six>=20221105

# Error monitoring — optional but strongly recommended for production runs
# Set SENTRY_DSN environment variable to enable
sentry-sdk>=2.0.0
